+++ src/app/(route)/customer/(route)/feedback/feedback-info/[feedbackId]/_components/FeedbackInfo.tsx
@@ -39,7 +39,6 @@ const FeedbackInfo = () => {
     isError: feedbackIsError,
   } = useGetFeedbackInfoData({ id });
 
-  console.log(adminRole);
   return (
     <>
       {/* 개선요청 상세 */}
