+++ src/app/(route)/customer/(route)/notice/notice-info/[noticeInfoId]/_components/templates/NoticeListContainer.tsx
@@ -1,37 +1,39 @@
 import React from "react";
-import NoticeItemSkeleton from "../NoticeInfoItemSkeleton";
-import Pagination from "@/app/(route)/mypage/_components/Pagination";
 import CustomerTalkToolbar from "@/app/(route)/customer/_components/ui/CustomerTalkToolbar";
-import EmptyItem from "@/app/(route)/customer/_components/common/EmptyItem";
-import NoticeItem from "../../../../_components/items/NoticeItem";
-import { cn } from "@/utils";
-import { NoticeContentType } from "@/app/(route)/customer/_types/NoticeItemType";
 import NoticeListBox from "../organisms/NoticeListBox";
+import useGetNoticeDataList from "@/_hooks/fetcher/customer/useGetNoticeDataList";
+import { noticeListConfig } from "@/app/(route)/customer/_types/noticeListConfig";
 
 interface NoticeListContainerProps {
-  isLoading: boolean;
-  isError: boolean;
-  noticeListData: any;
-  adminChecker: "ADMIN" | "USER" | undefined;
+  adminRole: "ADMIN" | "USER" | undefined;
   searchParams: any;
-  handlePageChange: (page: number) => void;
 }
 
 const NoticeListContainer = ({
-  isLoading,
-  isError,
-  noticeListData,
-  adminChecker,
+  adminRole,
   searchParams,
-  handlePageChange,
 }: NoticeListContainerProps) => {
+  const noticeOption: noticeListConfig = {
+    page: searchParams.get("page") ? Number(searchParams.get("page")) : 1,
+    size: 20,
+    searchType:
+      (searchParams.get("search_type") as noticeListConfig["searchType"]) || "",
+    search: searchParams.get("search") || "",
+  };
+
+  const {
+    data: noticeListData,
+    isLoading,
+    isError,
+  } = useGetNoticeDataList(noticeOption);
+
   return (
     <>
       {/* 툴바 */}
       <CustomerTalkToolbar
         showOptions={false}
         paginationData={noticeListData?.pageInfo}
-        adminChecker={adminChecker}
+        adminChecker={adminRole}
       />
 
       {/* 리스트 */}
@@ -40,7 +42,6 @@ const NoticeListContainer = ({
         isError={isError}
         noticeListData={noticeListData}
         searchParams={searchParams}
-        handlePageChange={handlePageChange}
       />
     </>
   );
